data shared with the GUI

+++ src/arduino/lib/Scheduling/SharableData.h
@@ -3,7 +3,7 @@
 
 struct sharableData{
     double temperature;
-    int fillPercentage;
+    double fillPercentage;
     int doorStatus;
 }
